@@ -1,11 +1,14 @@
 package io.mosip.biosdk.services;
 
 import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 
-@SpringBootApplication()
+@EnableAutoConfiguration(exclude = { DataSourceAutoConfiguration.class })
+@SpringBootApplication
 public class SdkApplication {
-    public static void main(String[] args) {
-        SpringApplication.run(SdkApplication.class, args);
-    }
+	public static void main(String[] args) {
+		SpringApplication.run(SdkApplication.class, args);
+	}
 }
